server/requirements.txt
rclpy
torch>=1.10.0
torchvision>=0.11.0
open3d>=0.13.0
opencv-python>=4.5.0
numpy>=1.19.0
scipy>=1.7.0
graspnetAPI
supervision>=0.16.0
hydra-core>=1.2.0
omegaconf>=2.2.0
pycocotools>=2.0.0
Pillow>=8.0.0
